requirements.txt
# Function dependencies, for example:
# package>=version
line_notify
datetime
pytz
google-cloud-bigquery
pandas

